rename utils to getEmails and make a separate utils file to fix confusions
