+++ emails/utils.ts
@@ -1,0 +1,15 @@
+export const formatHostNamesShort = (hosts: Array<{ name: string | null }>) => {
+  const namesWithValues = hosts
+    .filter((h) => h.name !== null)
+    .map((h) => h.name!);
+
+  if (namesWithValues.length === 0) {
+    return "the host";
+  }
+
+  if (namesWithValues.length <= 3) {
+    return namesWithValues.join(", ").replace(/, ([^,]*)$/, " and $1");
+  }
+
+  return `the hosts`;
+};
